refactor workspace code into own project
