+++ src/DigitalPreservation/DigitalPreservation.Workspace/CreateFolderResult.cs
@@ -1,4 +1,4 @@
-﻿namespace DigitalPreservation.UI.Workspace;
+﻿namespace DigitalPreservation.Workspace;
 
 public class CreateFolderResult
 {
