+++ src/DigitalPreservation/DigitalPreservation.Workspace/Requests/UploadFileToDeposit.cs
@@ -10,7 +10,7 @@ using MediatR;
 using Storage.Repository.Common;
 using Storage.Repository.Common.S3;
 
-namespace DigitalPreservation.UI.Features.Workspace;
+namespace DigitalPreservation.Workspace.Requests;
 
 public class UploadFileToDeposit(
     Uri s3Root,
