+++ src/DigitalPreservation/DigitalPreservation.sln
@@ -42,6 +42,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DigitalPreservation.XmlGen"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XmlGen.Tests", "XmlGen.Tests\XmlGen.Tests.csproj", "{CCA94442-F814-490B-9416-445F9AF8B060}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DigitalPreservation.Workspace", "DigitalPreservation.Workspace\DigitalPreservation.Workspace.csproj", "{C42868B6-06D5-45DD-BD1C-1BF2AEF30342}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -128,6 +130,10 @@ Global
 		{CCA94442-F814-490B-9416-445F9AF8B060}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CCA94442-F814-490B-9416-445F9AF8B060}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CCA94442-F814-490B-9416-445F9AF8B060}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C42868B6-06D5-45DD-BD1C-1BF2AEF30342}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C42868B6-06D5-45DD-BD1C-1BF2AEF30342}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C42868B6-06D5-45DD-BD1C-1BF2AEF30342}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C42868B6-06D5-45DD-BD1C-1BF2AEF30342}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{E045329F-268B-4E69-8C2A-82B2CBEB9014} = {2FF7269F-F433-41F9-84BF-EAA8CFCFFCF5}
