+++ src/DigitalPreservation/Preservation.API/Preservation.API.csproj
@@ -24,6 +24,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\DigitalPreservation.Workspace\DigitalPreservation.Workspace.csproj" />
       <ProjectReference Include="..\Storage.Client\Storage.Client.csproj" />
     </ItemGroup>
